requirements.txt
python-telegram-bot==13.15
Flask
requests
pandas
sqlite3
beautifulsoup4
